* [compiler/cmake] Added asan support, and updated MSVC build * [compiler/gpu] Added gpu arch specifier, and improved GPU codegen * [compiler] Supported more broadcast fusion, and improved elementwise codegen * [compiler/transformation] Supported more loop tag, and improved memory planning * [compiler/utils] Improved cloning, folding, and hoist * [compiler/CAT] Improved AIT supports in CAT * [compiler/mhlo] Added canonicalizers * [compiler] Fixed bugs * [compiler/E2E] Made default pipelines from affine to scf * [compiler] Bumped version to 1.1.0 * [frontend/onnx] Fixed readme * [frontend/tf] Improved TF frontend attributes * [frontend/torch] Bumped torch-mlir version * [runtime] Bumped version to 1.1.0 * [runtime] Added TF attribute support

+++ compiler/include/byteir/Transforms/LoopTag.h
@@ -29,7 +29,7 @@ class FuncOp;
 
 std::unique_ptr<OperationPass<func::FuncOp>>
 createLoopTagPass(llvm::StringRef anchorTag = "",
-                  const std::string &attachTag = "", unsigned depth = 1,
+                  const std::string &attachTag = "", int64_t depth = 1,
                   const std::string &loopType = "scf.for");
 
 } // namespace mlir
